chore(opapi): refactor return types parser
