+++ opapi/src/test.json
@@ -1,0 +1,116 @@
+{
+  "operations": {
+    "getFoo": {
+      "name": "getFoo",
+      "description": "Get a foo",
+      "method": "get",
+      "path": "//v1/foos/{id}",
+      "parameters": { "id": { "in": "path", "type": "string", "description": "Foo id" } },
+      "section": "foo",
+      "response": {
+        "description": "Foo information",
+        "schema": {
+          "type": "object",
+          "properties": { "foo": { "$ref": "#/components/schemas/Foo" } },
+          "required": ["foo"],
+          "title": "getFooResponse",
+          "additionalProperties": false
+        }
+      }
+    },
+    "postFoo": {
+      "name": "postFoo",
+      "description": "Post a foo",
+      "method": "post",
+      "path": "//v1/foos/{id}",
+      "parameters": { "id": { "in": "path", "type": "string", "description": "Foo id" } },
+      "requestBody": {
+        "description": "Foo information",
+        "schema": {
+          "type": "object",
+          "description": "Foo information",
+          "title": "postFooBody",
+          "additionalProperties": false
+        }
+      },
+      "section": "foo",
+      "response": {
+        "description": "Foo information",
+        "schema": {
+          "type": "object",
+          "properties": { "foo": { "$ref": "#/components/schemas/Foo" } },
+          "required": ["foo"],
+          "title": "postFooResponse",
+          "additionalProperties": false
+        }
+      }
+    },
+    "postBar": {
+      "name": "postBar",
+      "description": "Post a bar",
+      "method": "post",
+      "path": "//v1/bars",
+      "requestBody": {
+        "description": "Bar information",
+        "schema": {
+          "type": "object",
+          "properties": { "test": { "type": "string" } },
+          "required": ["test"],
+          "title": "postBarBody",
+          "additionalProperties": false
+        }
+      },
+      "section": "bar",
+      "response": {
+        "description": "Bar information",
+        "schema": {
+          "type": "object",
+          "properties": { "bar": { "$ref": "#/components/schemas/Bar" } },
+          "required": ["bar"],
+          "title": "postBarResponse",
+          "additionalProperties": false
+        }
+      },
+      "parameters": {}
+    }
+  },
+  "metadata": {
+    "title": "Test API",
+    "description": "Test API",
+    "server": "http://localhost:3000",
+    "version": "1.0.0",
+    "prefix": "/v1"
+  },
+  "errors": [
+    { "status": 500, "type": "Unknown", "description": "An unknown error occurred" },
+    { "status": 500, "type": "Internal", "description": "An internal error occurred" }
+  ],
+  "refs": {
+    "parameters": {},
+    "requestBodies": { "postFooBody": true, "postBarBody": true },
+    "responses": { "getFooResponse": true, "postFooResponse": true, "postBarResponse": true },
+    "schemas": { "Foo": true, "Bar": true }
+  },
+  "schemas": {
+    "Foo": {
+      "section": "foo",
+      "schema": {
+        "type": "object",
+        "properties": { "id": { "type": "string" }, "name": { "type": "string" } },
+        "required": ["id", "name"],
+        "additionalProperties": false
+      }
+    },
+    "Bar": { "section": "bar", "schema": { "type": "object", "additionalProperties": false } }
+  },
+  "sections": [
+    {
+      "title": "Foo",
+      "description": "Foo section",
+      "name": "foo",
+      "operations": ["getFoo", "postFoo"],
+      "schema": "Foo"
+    },
+    { "title": "Bar", "description": "Bar section", "name": "bar", "operations": ["postBar"], "schema": "Bar" }
+  ]
+}
