+++ opapi/src/{"openapi":"3.0.json
@@ -1,0 +1,129 @@
+{
+  "openapi": "3.0.0",
+  "servers": [{ "url": "http://localhost:3000" }],
+  "info": { "title": "Test API", "description": "Test API", "version": "1.0.0" },
+  "paths": {
+    "//v1/foos/{id}": {
+      "get": {
+        "operationId": "getFoo",
+        "description": "Get a foo",
+        "parameters": [
+          { "name": "id", "in": "path", "description": "Foo id", "required": true, "schema": { "type": "string" } }
+        ],
+        "responses": {
+          "200": { "$ref": "#/components/responses/getFooResponse" },
+          "default": { "$ref": "#/components/responses/getFooResponse" }
+        }
+      },
+      "post": {
+        "operationId": "postFoo",
+        "description": "Post a foo",
+        "parameters": [
+          { "name": "id", "in": "path", "description": "Foo id", "required": true, "schema": { "type": "string" } }
+        ],
+        "responses": {
+          "200": { "$ref": "#/components/responses/postFooResponse" },
+          "default": { "$ref": "#/components/responses/postFooResponse" }
+        },
+        "requestBody": { "$ref": "#/components/requestBodies/postFooBody" }
+      }
+    },
+    "//v1/bars": {
+      "post": {
+        "operationId": "postBar",
+        "description": "Post a bar",
+        "parameters": [],
+        "responses": {
+          "200": { "$ref": "#/components/responses/postBarResponse" },
+          "default": { "$ref": "#/components/responses/postBarResponse" }
+        },
+        "requestBody": { "$ref": "#/components/requestBodies/postBarBody" }
+      }
+    }
+  },
+  "components": {
+    "schemas": {
+      "Foo": {
+        "type": "object",
+        "properties": { "id": { "type": "string" }, "name": { "type": "string" } },
+        "required": ["id", "name"],
+        "additionalProperties": false
+      },
+      "Bar": { "type": "object", "additionalProperties": false }
+    },
+    "responses": {
+      "getFooResponse": {
+        "description": "Foo information",
+        "content": {
+          "application/json": {
+            "schema": {
+              "type": "object",
+              "properties": { "foo": { "$ref": "#/components/schemas/Foo" } },
+              "required": ["foo"],
+              "title": "getFooResponse",
+              "additionalProperties": false
+            }
+          }
+        }
+      },
+      "postFooResponse": {
+        "description": "Foo information",
+        "content": {
+          "application/json": {
+            "schema": {
+              "type": "object",
+              "properties": { "foo": { "$ref": "#/components/schemas/Foo" } },
+              "required": ["foo"],
+              "title": "postFooResponse",
+              "additionalProperties": false
+            }
+          }
+        }
+      },
+      "postBarResponse": {
+        "description": "Bar information",
+        "content": {
+          "application/json": {
+            "schema": {
+              "type": "object",
+              "properties": { "bar": { "$ref": "#/components/schemas/Bar" } },
+              "required": ["bar"],
+              "title": "postBarResponse",
+              "additionalProperties": false
+            }
+          }
+        }
+      }
+    },
+    "requestBodies": {
+      "postFooBody": {
+        "description": "Foo information",
+        "content": {
+          "application/json": {
+            "schema": {
+              "type": "object",
+              "description": "Foo information",
+              "title": "postFooBody",
+              "additionalProperties": false
+            }
+          }
+        }
+      },
+      "postBarBody": {
+        "description": "Bar information",
+        "content": {
+          "application/json": {
+            "schema": {
+              "type": "object",
+              "properties": { "test": { "type": "string" } },
+              "required": ["test"],
+              "title": "postBarBody",
+              "additionalProperties": false
+            }
+          }
+        }
+      }
+    },
+    "parameters": {}
+  }
+}
